P3.2: split headless vs display tests
